nemukerja/requirements.txt
Flask
Flask-WTF
Flask-Login
Flask-Bcrypt
Flask-SQLAlchemy
PyMySQL
python-dotenv
Flask-Migrate
email-validator
Flask-Babel